Fixed `PageTestCase` due to different pages
All pages are tested now (with success); the only page *not* tested is
`blog` since it requires Google API authentication, which I'm not
willing to copy over from `production.py` to `test.py` due to privacy
issues.

It could be possible to create something like `test-credentials.py` and
exclude it from GIT, but I don't feel comfortable with that either,
since it would add a *second* of the grid configuration file like
*production.py*.

So I decided to simple drop the test for the `blog` page.